# internal/http/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory Back API
  version: 0.2.0
servers:
  - url: /api/v1
...
paths:
...
  /products:
    get:
      tags: [products]
      summary: Search products
      description: Non-admin users must provide group_id and only see that group's products.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: group_id
          schema: { type: string, format: uuid }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, example: active }
        - in: query
          name: paid
          schema: { type: string, enum: ["true","false"] }
        - in: query
          name: min_price
          schema: { type: number, format: double }
        - in: query
          name: max_price
          schema: { type: number, format: double }
        - in: query
          name: from_entry
          schema: { type: string, format: date-time }
        - in: query
          name: to_entry
          schema: { type: string, format: date-time }
        - in: query
          name: sort
          schema:
            type: string
            enum: [created_at_desc, created_at_asc, price_asc, price_desc, entry_date_asc, entry_date_desc]
        - in: query
          name: limit
          schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Search result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"
        "403":
          description: Forbidden

    post:
      tags: [products]
      summary: Create product
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "403":
          description: Forbidden
  /products/{id}:
    get:
      tags: [products]
      summary: Get product by id
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "403":
          description: Forbidden
        "404":
          description: Not found

    put:
      tags: [products]
      summary: Update product by id
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "403":
          description: Forbidden
        "404":
          description: Not found

    delete:
      tags: [products]
      summary: Delete product by id
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Contact:
      type: object
      required: [first_name, last_name, phone_number]
      properties:
        first_name: { type: string }
        last_name: { type: string }
        phone_number: { type: string }
    Image:
      type: object
      required: [image_url]
      properties:
        image_url: { type: string }
        position: { type: integer, default: 0 }
    ProductCreate:
      type: object
      required: [group_id, name, price, status, entry_date, paid]
      properties:
        group_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        price: { type: number, format: double }
        paid: { type: boolean }
        status: { type: string, enum: [draft, active, archived] }
        entry_date: { type: string, format: date-time }
        exit_date: { type: string, format: date-time, nullable: true }
        observations: { type: string }
        contact:
          $ref: "#/components/schemas/Contact"
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
    ProductUpdate:
      type: object
      required: [name, price, status, entry_date, paid]
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number, format: double }
        paid: { type: boolean }
        status: { type: string, enum: [draft, active, archived] }
        entry_date: { type: string, format: date-time }
        exit_date: { type: string, format: date-time, nullable: true }
        observations: { type: string }
    Product:
      allOf:
        - $ref: "#/components/schemas/ProductCreate"
        - type: object
          properties:
            id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    SearchResult:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
...